+++ R/AAA_registry.R
@@ -10,13 +10,15 @@ recommenderRegistry$set_field("fun", type = "function",
     is_key = FALSE)
 recommenderRegistry$set_field("description", type = "character",
     is_key = FALSE)
+recommenderRegistry$set_field("reference", type = "character",
+    is_key = FALSE)
 recommenderRegistry$set_field("parameters", type = "list",
     is_key = FALSE)
 
 print.recommender_method <- function(x, ...) {
   with(x, {
-    writeLines(sprintf("Recommender method: %s\nDescription: %s",
-                       method, description))
+    writeLines(sprintf("Recommender method: %s for %s\nDescription: %s\nReference: %s",
+                       method, dataType, description, reference))
 
     if(is.list(parameters)) {
       writeLines("Parameters:")